get bearer tokens working
