+++ client/src/app/auth/tabs/dashboard/accounts/delete-account.tsx
@@ -4,8 +4,7 @@ import { Button } from '@/components/ui/button';
 import { Checkbox } from '@/components/ui/checkbox';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 import { useToast } from '@/components/ui/use-toast';
-import { deleteAccount } from '@/lib/accounts';
-import { translateAxiosError } from '@/lib/request';
+import { translateAxiosError } from '@/lib/requests';
 import { TrashIcon } from '@radix-ui/react-icons';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { type AxiosError } from 'axios';
@@ -18,14 +17,17 @@ interface DeleteAccountProps {
 const DeleteAccount = (props: DeleteAccountProps): JSX.Element => {
   const [deleteTransactionsValue, setDeleteTransactionsValue] = React.useState(false);
 
-  const { accessToken } = React.useContext<any>(AuthContext);
+  const { request } = React.useContext<any>(AuthContext);
 
   const { toast } = useToast();
   const queryClient = useQueryClient();
   const doDeleteAccount = useMutation({
-    mutationFn: async (deleteTransactions: boolean) => {
-      return await deleteAccount(accessToken, props.accountId, deleteTransactions);
-    },
+    mutationFn: async (deleteTransactions: boolean) =>
+      await request({
+        url: '/api/account',
+        method: 'DELETE',
+        params: { guid: props.accountId, deleteTransactions },
+      }),
     onSuccess: async () => {
       await queryClient.invalidateQueries({ queryKey: ['accounts'] });
     },
