+++ client/src/app/auth/tabs/dashboard/sync-account-button.tsx
@@ -1,21 +1,18 @@
 import { AuthContext } from '@/components/auth-provider';
 import ResponsiveButton from '@/components/responsive-button';
 import { useToast } from '@/components/ui/use-toast';
-import { translateAxiosError } from '@/lib/request';
-import { doSync } from '@/lib/user';
+import { translateAxiosError } from '@/lib/requests';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { type AxiosError } from 'axios';
 import React from 'react';
 
 const SyncAccountButton = (): JSX.Element => {
-  const { accessToken } = React.useContext<any>(AuthContext);
+  const { request } = React.useContext<any>(AuthContext);
 
   const { toast } = useToast();
   const queryClient = useQueryClient();
   const doSyncMutation = useMutation({
-    mutationFn: async () => {
-      return await doSync(accessToken);
-    },
+    mutationFn: async () => await request({ url: '/api/simplefin/sync', method: 'GET' }),
     onSuccess: async () => {
       await queryClient.invalidateQueries({ queryKey: ['transactions'] });
       await queryClient.invalidateQueries({ queryKey: ['accounts'] });
