+++ client/src/components/auth-route.tsx
@@ -11,7 +11,7 @@ const AuthRoute = ({ children }: { children: any }): JSX.Element => {
   if (loading) {
     return <PageLoading />;
   }
-  if (accessToken.length > 0) {
+  if (accessToken != null) {
     return children;
   }
   return <Navigate to="/" />;
