+++ client/src/lib/goals.ts
@@ -1,44 +1,7 @@
-import { AxiosResponse } from 'axios';
-import request from './request';
-import { Goal, NewGoal } from '@/types/goal';
+import { Goal } from '@/types/goal';
 import { getTransactionsForMonth } from './transactions';
 import { getMonthsUntilDate } from './utils';
 import { sumAccountsTotalBalance } from './accounts';
-
-export const getGoals = async (accessToken: string): Promise<AxiosResponse> =>
-  await request(
-    {
-      url: '/api/goal',
-      method: 'GET',
-    },
-    accessToken
-  );
-
-export const addGoal = async (
-  accessToken: string,
-  newGoal: NewGoal
-): Promise<AxiosResponse> =>
-  await request(
-    {
-      url: '/api/goal',
-      method: 'POST',
-      data: newGoal,
-    },
-    accessToken
-  );
-
-export const deleteGoal = async (
-  accessToken: string,
-  guid: string
-): Promise<AxiosResponse> =>
-  await request(
-    {
-      url: '/api/goal',
-      method: 'DELETE',
-      params: { guid },
-    },
-    accessToken
-  );
 
 export const getGoalTargetAmount = (amount: number, initialAmount: number): number => {
   if (initialAmount < 0) {
